add api endpoint

+++ src/main/java/com/thesis/bikerental/portfolio/receipt/service/ReceiptRepository.java
@@ -3,5 +3,5 @@ package com.thesis.bikerental.portfolio.receipt.service;
 import com.thesis.bikerental.portfolio.receipt.domain.Receipt;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface ReceiptRepository extends JpaRepository<Receipt, Integer> {
+public interface ReceiptRepository extends JpaRepository<Receipt, Long> {
 }
